Product Index first part

+++ app/Http/Controllers/ProductController.php
@@ -7,8 +7,12 @@ use Illuminate\Http\Request;
 use App\Models\Product;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
+use App\Http\Resources\CategoryResource;
+use App\Http\Resources\MeasureUnitResource;
 use App\Http\Resources\ProductResource;
 use App\Imports\ProductsImport;
+use App\Models\Category;
+use App\Models\MeasureUnit;
 use Inertia\Inertia;
 use Maatwebsite\Excel\Facades\Excel;
 
@@ -28,18 +32,22 @@ class ProductController extends Controller
     }
 
     public function index() {
-        $products = Product::all();
+        $products = Product::with(['category', 'measureUnit', 'retailMeasureUnit'])->get();
+        $categories = Category::all();
+        $measureUnits = MeasureUnit::all();
         return Inertia::render('Products/Index', [
-            'products' => $products
+            'products' => ProductResource::collection($products),
+            'categories' => CategoryResource::collection($categories),
+            'measureUnits' => MeasureUnitResource::collection($measureUnits),
         ]);
     }
 
-    public function show($id) {
+    /* public function show($id) {
         $product = Product::findOrFail($id);
         return Inertia::render('Products/Show', [
             'product' => $product
         ]);
-    }
+    } */
 
     public function search(Request $request) {
         $q = trim((string) $request->query('q'));
@@ -77,6 +85,11 @@ class ProductController extends Controller
 
     public function destroy($id) {
         $product = Product::findOrFail($id);
+        $relatedPurchaseItemsCount = $product->purchaseItems()->count();
+        $relatedSaleItemsCount = $product->saleItems()->count();
+        if ($relatedPurchaseItemsCount > 0 || $relatedSaleItemsCount > 0) {
+            return redirect()->route('products.index')->with('error', 'No se puede eliminar el producto porque tiene registros relacionados de compras o ventas.');
+        }
         $product->delete();
         return redirect()->route('products.index')->with('success', 'Producto eliminado exitosamente.');
     }
